requirements.txt
asgiref==3.3.0
autopep8==1.5.4
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.3
django-background-tasks==1.2.5
django-braces==1.14.0
django-compat==1.0.15
django-filter==2.4.0
django-oauth-toolkit==1.3.3
django-rest-framework-social-oauth2==1.1.0
django-storages==1.10.1
djangorestframework==3.12.2
google==3.0.0
google-api-core==1.23.0
google-auth==1.23.0
google-cloud==0.34.0
google-cloud-core==1.4.4
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-oauth==1.0.1
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
Markdown==3.3.3
oauthlib==3.1.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.1
whitenoise==5.2.0
